+++ package.json
@@ -46,6 +46,7 @@
     "compression": "^1.7.3",
     "connect-session-sequelize": "^6.0.0",
     "css-loader": "^3.0.0",
+    "dotenv": "^8.0.0",
     "express": "^4.16.4",
     "express-session": "^1.15.1",
     "history": "^4.9.0",
@@ -59,11 +60,14 @@
     "react-dom": "^16.8.6",
     "react-redux": "^7.0.1",
     "react-router-dom": "^5.0.0",
+    "react-stripe-checkout": "^2.6.3",
+    "react-toastify": "^5.2.1",
     "redux": "^4.0.1",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
     "sequelize": "^5.3.1",
     "socket.io": "^2.2.0",
+    "stripe": "^7.1.0",
     "style-loader": "^0.23.1"
   },
   "devDependencies": {
