+++ server/api/users.js
@@ -1,6 +1,9 @@
 const router = require('express').Router()
 const {User, Order} = require('../db/models')
 module.exports = router
+const stripe = require('stripe')('sk_test_tjWlFFD447mw8htkrmXiorw3009wcWw6Wt')
+
+const uuid = require('uuid/v4')
 
 router.get('/', async (req, res, next) => {
   try {
@@ -80,4 +83,52 @@ router.put('/address', async (req, res, next) => {
   } catch (error) {
     next(error)
   }
+})
+
+router.get('/orders/checkout', async (req, res) => {
+  res.send(`Add your Stripe Secret Key to the .require('stripe" statement!`)
+})
+
+router.post('/checkout', async (req, res, next) => {
+  console.log('Request:', req.body)
+  let error
+  let status
+  try {
+    const {product, token} = req.body
+    console.log('this is the req.body:', req.body)
+    const customer = await stripe.customers.create({
+      email: token.email,
+      source: token.id
+    })
+    const idempotency_key = uuid()
+    const charge = await stripe.charges.create(
+      {
+        amount: product.price * 100,
+        currency: 'usd',
+        customer: customer.id,
+        receipt_email: token.email,
+        description: `Purchased the ${product.name}`,
+        shipping: {
+          name: token.card.name,
+          address: {
+            line1: token.card.address_line1,
+            line2: token.card.address_line2,
+            city: token.card.address_city,
+            country: token.card.address_country,
+            postal_code: token.card.address_zip
+          }
+        }
+      },
+      {
+        idempotency_key
+      }
+    )
+    console.log('Charge:', {charge})
+    status = 'success'
+  } catch (err) {
+    console.error('Error:', err)
+    status = 'failure'
+  }
+
+  res.json({error, status})
 })
